backend/requirements.txt
lask
google-generativeai
python-dotenv
flask-cors